Bug Fixes + Better Output
Fixed a Big issue where IP:Port combinations in a Range, were not scanning from the beginning of the Paths array if credentials were found. Now working better.

Reduced the Timeout Delay from 60s to 30s (needs some testing but should be better)

Added loads more output to make it easier to read, including outputting found credentials as they are found, and creating a full valid RTSP path for easy copying.

Still to do is Digest Auth, I know how to do it, but need a way to test it correctly.